pyproject.toml
[project]
name = "workiq-agents-demo"
version = "0.1.0"
description = "A Python app that reads meeting transcripts via Work IQ, plans tasks with GitHub Copilot SDK, creates repos/issues, and assigns coding agents."
requires-python = ">=3.11"
dependencies = [
    "github-copilot-sdk>=0.1.0",
    "PyGithub>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.scripts]
workiq-demo = "main:cli"
